+++ wall.cpp
@@ -1,6 +1,6 @@
 #include "wall.h"
 
-Wall::Wall(int _x, int _y)
+Wall::Wall(float _x, float _y)
     : Obstacle(_x,_y)
 {
